Rename `neo4j` package to `neo4j_importer` to fix import conflict
