+++ kuwala-core/data-importer/main.py
@@ -2,7 +2,7 @@ import sys
 
 sys.path.insert(0, './src')
 
-import src.neo4j.PipelineImporter as PipelineImporter
+import src.neo4j_importer.PipelineImporter as PipelineImporter
 
 if __name__ == '__main__':
     PipelineImporter.import_pipelines()
