+++ kuwala-core/data-importer/src/neo4j_importer/PoiOSMImporter.py
@@ -1,8 +1,8 @@
 import h3
 import json
 import os
-import src.neo4j.Neo4jConnection as Neo4jConnection
-import src.neo4j.PipelineImporter as PipelineImporter
+import src.neo4j_importer.Neo4jConnection as Neo4jConnection
+import src.neo4j_importer.PipelineImporter as PipelineImporter
 from pyspark.sql.functions import flatten, lit
